refactor(todoorder): projection core → internal/ shared package (FMT-33 official pattern) + rebuild slice L0
- move projection Service/store/handlers/Query/Rebuild to cells/ordercell/internal/orderprojection (configreader precedent for FMT-33 public/internal slice sharing)
- orderprojection slice → thin alias wrapper (type Service = internalproj.Service); orderprojectionrebuild imports internal pkg → eliminates slice→slice import
- orderprojectionrebuild consistencyLevel L3→L0 (local in-memory replay, no tx/event/cross-cell)
- regenerate slice_gen + assembly boundary; drop alias slice's required-deps gen
