+++ examples/todoorder/cells/ordercell/internal/orderprojection/service_required_gen.go
